Web Application for testing of cloudformation
In "css" folder will be found main and bootstrap css file; main.css is
used for new css style.
In "js" folder will be found main, jquery and bootstrap js file. For
this project, there have no used of main.js file.
In "img" floder will be found the images for this application.
And, the "index.html" is the main web application file.